feat: extract map utilities
