+++ index.html
@@ -45,7 +45,7 @@
   <script defer src="js/log_test_summary.js"></script>
   <script defer src="js/chart.js"></script>
   <script defer src="js/update_stats.js"></script>
-  <script defer src="js/map.js"></script>
+  <script type="module" src="js/map.js"></script>
     <script defer src="js/map_zoom.js"></script>
     <script defer src="js/find_admin_unit.js"></script>
     <script defer src="js/hromada_change_announcer.js"></script>
@@ -62,7 +62,7 @@
   <script defer src="js/replace_spaces_with_underscore.js"></script>
   <script defer src="js/download_CSV.js"></script>
   <script defer src="js/download_KML.js"></script>
-  <script defer src="js/download_HTML.js"></script>
+  <script type="module" src="js/download_HTML.js"></script>
   <script defer src="js/download_chart.js"></script>
   <script defer src="js/clear_data.js"></script>
   <script defer src="js/update_UI.js"></script>
